backend/requirements.txt
# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# Validation
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0

# HTTP
httpx==0.27.2

# Vector DB
qdrant-client==1.11.3

# AI / ML
anthropic==0.34.2
google-cloud-vision==3.7.4
google-cloud-aiplatform==1.68.0

# Storage
boto3==1.35.24

# Utils
python-dotenv==1.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
